# frontend/api-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Library API",
    "description": "API для управления библиотекой",
    "version": "v1"
  },
  "paths": {
...
    "/users/{userId}/my-books": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Получить список книг, взятых текущим пользователем",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageNo",
            "in": "query",
            "description": "Номер страницы (опционально)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "Количество элементов на странице (опционально)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список успешно получен",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookLendingDTOIEnumerablePaginationListModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookLendingDTOIEnumerablePaginationListModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookLendingDTOIEnumerablePaginationListModel"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/books/catalog": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Получить список книг по фильтру",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "description": "Строка поиска (опционально)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genreId",
            "in": "query",
            "description": "ID жанра (опционально)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AuthorId",
            "in": "query",
            "description": "ID автора (опционально)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageNo",
            "in": "query",
            "description": "Номер страницы (опционально)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "Количество элементов на странице (опционально)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список успешно получен",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookCatalogDTOResponseData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookCatalogDTOResponseData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookCatalogDTOResponseData"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Получить книгу по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID книги",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Книга успешно получена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailsDTO"
                }
              }
            }
          },
          "404": {
            "description": "Книга не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/books/{isbn}": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Получить книгу по ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "description": "ISBN книги",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Книга успешно получена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailsDTO"
                }
              }
            }
          },
          "404": {
            "description": "Книга не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/users/{userId}/books/{bookId}/lend": {
      "post": {
        "tags": [
          "Book"
        ],
        "summary": "Выдать книгу",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "ID книги",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Книга успешно взята"
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Книга не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Книга уже взята",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/users/{userId}/books/{bookId}/return": {
      "post": {
        "tags": [
          "Book"
        ],
        "summary": "Вернуть книгу",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "bookId",
            "in": "path",
            "description": "ID книги",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Книга успешно возвращена"
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Книга не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
...
    "/books/{id}/upload-cover": {
      "post": {
        "tags": [
          "Book"
        ],
        "summary": "Загрузить изображение книги",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID книги",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "Изображение обложки книги",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "image": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Изображение успешно загружено"
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Книга не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/books/{id}/remove-cover": {
      "post": {
        "tags": [
          "Book"
        ],
        "summary": "Установить изображение книги по умолчанию",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID книги",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Изображение успешно установлено"
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Книга не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BookCatalogDTOResponseData": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BookCatalogDTO"
          },
          "success": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookDetailsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "genreId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "genre": {
            "$ref": "#/components/schemas/Genre"
          },
          "isbn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BookLendingDTOIEnumerablePaginationListModel": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BookLendingDTO"
              }
            },
            "nullable": true
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}